requirements.txt
numpy==2.0.1
pillow==11.0.0
pytorch==2.5.1
timm==1.0.20
tqdm==4.67.1
transformers==4.57.1
tensorboard==2.19.0